requirements.txt
opencv-python
mediapipe
numpy
pillow
scikit-learn
pandas
joblib
